tripDiary: extract the segment section's isSectionVisible function
This function has the same logic as the one for the visited places
section, so it is extracted in a file in
`questionnaire/sections/tripDiary` folder. The condition is changed to
use the `shouldShowTripsAndPlacesSections` OD helper function
